Default to doing DNS lookup through proxy server
This __init__.py is has the same version in the header.  However this
was submitted on the 14 Jul and was not included in the 28 Jul release
bundled with GAM.

Relevant for GAM user operating through a proxy.  Does the following

- Resolve an issue where proxy information by default was not resolving
DNS through a proxy server (unlike the `socks` library), so users
getting proxy information from an environment variable
(HTTP_PROXY/HTTPS_PROXY)
with no external DNS could not connect to hosts with DNS names.

- Resolve a potential issue where if the proxy server was a different
address family to the destination host and there was functional DNS, may
not work.  For example, using a IPv4-only proxy server to connect to an
IPv6-only host via an IPv4-only internal network.